Move where the enum is to hopefully fix tests
